requirements.txt
anthropic>=0.90.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.50
